feat: EntityLoader metaData 사용
