+++ src/test/java/persistence/entity/CustomJpaRepositoryTest.java
@@ -14,12 +14,12 @@ import java.sql.SQLException;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class CustomJpaRepositoryTest extends DatabaseTest {
-    private EntityLoader queryBuilder;
+    private QueryBuilder queryBuilder;
 
     @BeforeEach
     public void beforeEach() throws SQLException {
         super.beforeEach();
-        queryBuilder = new EntityLoader(new H2SelectQueryBuilder(), new H2DeleteQueryBuilder(), new H2InsertQueryBuilder(), new H2UpdateQueryBuilder(), jdbcTemplate);
+        queryBuilder = new QueryBuilder(new H2SelectQueryBuilder(), new H2DeleteQueryBuilder(), new H2InsertQueryBuilder(), new H2UpdateQueryBuilder(), jdbcTemplate);
     }
 
     @Test
